chore: Rename test blocks file for move mode
