+++ packages/blockly/tests/mocha/test_helpers/move_test_blocks.js
